KsModel improved

+++ src/KsModel/Struct_Ksmodel.jl
@@ -29,16 +29,16 @@ module ksModel
 		τ₂Mac_Max :: Vector{Float64}
 		τ₃Mac_Max :: Vector{Float64}
 
-		Nse_τ     :: 	Vector{Float64}
-		Rmse_τ    :: 	Vector{Float64}
-		Wilmot_τ  :: 	Vector{Float64}
+		Nse_τ     :: Vector{Float64}
+		Rmse_τ    :: Vector{Float64}
+		Wilmot_τ  :: Vector{Float64}
 	end # mutable struct KSMODEL
 
 	
 	# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 	#		FUNCTION : STRUCT_KSMODEL
 	# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-		function STRUCT_KSMODEL(;Nτ_Layer = 2)
+		function STRUCT_KSMODEL(;Nτ_Layer = 2::Int64)
 			τ₁       = fill(0.0::Float64, Nτ_Layer)
 			τ₂       = fill(0.0::Float64, Nτ_Layer)
 			τ₃       = fill(0.0::Float64, Nτ_Layer)
@@ -67,7 +67,7 @@ module ksModel
 			Rmse_τ = fill(0.0::Float64, Nτ_Layer)
 			Wilmot_τ = fill(0.0::Float64, Nτ_Layer)
 
-			ksmodelτ = KSMODELτ(τ₁, τ₂, τ₃, τ₄, τ₁Mac,τ₂Mac,τ₃Mac,τ₁_Min,τ₂_Min,τ₃_Min,  τ₄_Min, τ₁Mac_Min,τ₂Mac_Min,τ₃Mac_Min,τ₁_Max,τ₂_Max,τ₃_Max,τ₄_Max,τ₁Mac_Max,τ₂Mac_Max,τ₃Mac_Max,Nse_τ,Rmse_τ, Wilmot_τ)
+			ksmodelτ = KSMODELτ(τ₁, τ₂, τ₃, τ₄, τ₁Mac, τ₂Mac, τ₃Mac, τ₁_Min, τ₂_Min, τ₃_Min, τ₄_Min, τ₁Mac_Min, τ₂Mac_Min, τ₃Mac_Min, τ₁_Max, τ₂_Max, τ₃_Max, τ₄_Max, τ₁Mac_Max,τ₂Mac_Max, τ₃Mac_Max, Nse_τ, Rmse_τ, Wilmot_τ)
 
 		return ksmodelτ 
 		end  # function: STRUCT_KSMODEL
